+++ demos/chinook/src/main/java/is/codion/demos/chinook/model/InvoiceEditModel.java
@@ -20,7 +20,8 @@ package is.codion.demos.chinook.model;
 
 import is.codion.framework.db.EntityConnectionProvider;
 import is.codion.framework.domain.entity.Entity;
-import is.codion.framework.model.EntityEditor.EditorValue.Propagator;
+import is.codion.framework.domain.entity.attribute.Attribute;
+import is.codion.framework.model.EntityEditor.EditorValue;
 import is.codion.swing.framework.model.SwingEntityEditModel;
 
 import org.jspecify.annotations.Nullable;
@@ -32,27 +33,20 @@ public final class InvoiceEditModel extends SwingEntityEditModel {
 
 	public InvoiceEditModel(EntityConnectionProvider connectionProvider) {
 		super(Invoice.TYPE, connectionProvider);
+		EditorValue<Entity> customer = editor().value(Invoice.CUSTOMER_FK);
 		// By default, foreign key values persist when the model
 		// is cleared, here we disable that for CUSTOMER_FK
-		editor().value(Invoice.CUSTOMER_FK).persist().set(false);
+		customer.persist().set(false);
 		// We populate the invoice address fields with
 		// the customer address when the customer is edited
-		editor().value(Invoice.CUSTOMER_FK).propagate(new SetBillingAddress());
+		customer.propagate(Invoice.BILLINGADDRESS, cust -> valueOrNull(cust, Customer.ADDRESS));
+		customer.propagate(Invoice.BILLINGCITY, cust -> valueOrNull(cust, Customer.CITY));
+		customer.propagate(Invoice.BILLINGPOSTALCODE, cust -> valueOrNull(cust, Customer.POSTALCODE));
+		customer.propagate(Invoice.BILLINGSTATE, cust -> valueOrNull(cust, Customer.STATE));
+		customer.propagate(Invoice.BILLINGCOUNTRY, cust -> valueOrNull(cust, Customer.COUNTRY));
 	}
 
-	// Implement a Propagator, which updates the billing address when the
-	// invoice customer is changed. The value propagation is applied each time
-	// editing happens via the underlying editor, or when triggered externally,
-	// such as in a table, via an editable table cell or multi item editing
-	private static final class SetBillingAddress implements Propagator<Entity> {
-
-		@Override
-		public void propagate(@Nullable Entity customer, Setter setter) {
-			setter.set(Invoice.BILLINGADDRESS, customer == null ? null : customer.get(Customer.ADDRESS));
-			setter.set(Invoice.BILLINGCITY, customer == null ? null : customer.get(Customer.CITY));
-			setter.set(Invoice.BILLINGPOSTALCODE, customer == null ? null : customer.get(Customer.POSTALCODE));
-			setter.set(Invoice.BILLINGSTATE, customer == null ? null : customer.get(Customer.STATE));
-			setter.set(Invoice.BILLINGCOUNTRY, customer == null ? null : customer.get(Customer.COUNTRY));
-		}
+	private static @Nullable <T> T valueOrNull(Entity customer, Attribute<T> attribute) {
+		return customer == null ? null : customer.get(attribute);
 	}
 }