+++ changelog.md
@@ -30,6 +30,7 @@ Codion Change Log
 - EntityEditor.publishPersistenceEvents() moved to Settings, minor cleanup.
 - EntityTableModel.refreshTask() removed.
 - EntityTableModel, EntityModel.editor() added.
+- EntityEditor.EditorValue.Propagator removed. 
 ### is.codion.swing.framework.ui
 - TabbedDetailLayout.TabbedDetailController.selectParentTabs() bug fixed, now uses the correct indexOfComponent() instead of indexOfTabComponent().
 - TabbedApplicationLayout and TabbedDetailLayout now initialize panels using prepareComponent(), panels now contain focusable components when shown.